# server/swagger/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Try out API to Common Data Elements"
  version: "0.0.1"
  title: "CDE API"
host: "localhost:3001"
basePath: "/"
schemes:
  - "http"
produces:
  - "application/json"
paths:
...
  /api/form/{tinyId}:
    get:
      tags:
        - "Form"
      summary: "CDE Form"
      description: "Form template for data collection.\n"
      parameters:
        - name: "tinyId"
          in: "path"
          description: "identifier"
          required: true
          type: "string"
      responses:
        200:
          description: "CDE Form"
          schema:
            $ref: "#/definitions/Form"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "CDE"
  /api/form/{tinyId}/version/{version}:
    get:
      tags:
        - "Form"
      summary: "CDE Form"
      description: "Form template for data collection.\n"
      parameters:
        - name: "tinyId"
          in: "path"
          description: "identifier"
          required: true
          type: "string"
        - name: "version"
          in: "path"
          description: "version"
          required: true
          type: "string"
      responses:
        200:
          description: "CDE Form"
          schema:
            $ref: "#/definitions/Form"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "CDE"
  /api/form/search:
    post:
      tags:
        - "Form"
      summary: "Form document search"
      description: "Full-text search for Form documents"
      consumes:
        - application/json
      parameters:
        - in: body
          name: "Search Options"
          schema:
            $ref: "#/definitions/SearchApiOptions"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SearchDocumentResponseForm"
definitions:
...
  Form:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "57112ebf3245ae3033584c5a"
      nihEndorsed:
        type: "boolean"
        example: "true"
      tinyId:
        type: "string"
        example: "QF7dUNkbNEv"
      version:
        type: "string"
        example: "1.0"
      archived:
        type: "boolean"
      changeNote:
        type: "string"
        example: "load NCI on January 15 2020"
      comments:
        $ref: "#/definitions/Comments"
      copyright:
        type: "object"
        properties:
          authority:
            type: "string"
            example: "copyright NCI 2015"
          text:
            type: "string"
            example: "to be printed with permission only"
      created:
        type: "string"
        format: "date-time"
      createdBy:
        $ref: "#/definitions/UserRef"
      elementType:
        type: "string"
        example: "form"
      imported:
        type: "string"
        format: "date-time"
      isCopyrighted:
        type: "boolean"
      lastMigrationScript:
        type: "string"
        example: "load NINDS on January 15 2020"
      noRenderAllowed:
        type: "boolean"
      origin:
        type: "string"
        example: "NCISTD: NCI CRF Harmonization and Standards Initiative"
      registrationState:
        $ref: "#/definitions/RegistrationState"
      sources:
        $ref: "#/definitions/Source"
      stewardOrg:
        type: "object"
        properties:
          name:
            type: "string"
            example: "NINDS"
      updated:
        type: "string"
        format: "date-time"
      updatedBy:
        $ref: "#/definitions/UserRef"
      formElements:
        type: "array"
        items:
          type: "object"
          properties:
            elementType:
              type: "string"
              example: "section"
            instructions:
              $ref: "#/definitions/Instructions"
            inForm:
              type: "object"
              properties:
                form:
                  type: "object"
                  properties:
                    name:
                      type: "string"
                      example: "SubForm 1"
                    tinyId:
                      type: "string"
                      example: "QF7dUNkbNEv"
                    version:
                      type: "string"
                      example: "1.0"
            label:
              type: "string"
              example: "Clinical Section"
            question:
              type: "object"
              properties:
                answers:
                  $ref: "#/definitions/PermissibleValues"
                cde:
                  type: "object"
                  properties:
                    ids:
                      $ref: "#/definitions/Ids"
                    name:
                      type: "string"
                      example: "Starting Date"
                    permissibleValues:
                      $ref: "#/definitions/PermissibleValues"
                    tinyId:
                      type: "string"
                      example: "QF7dUNkbNEv"
                    version:
                      type: "string"
                      example: "1.0"
                datatype:
                  type: "string"
                  example: "Date"
                datatypeDate:
                  type: "object"
                  description: "object appearing only on Date datatypes"
                  properties:
                    format:
                      type: "string"
                      example: "YY-MM-DD"
                datatypeDynamicCodeList:
                  type: "object"
                  description: "object appearing only on Code datatypes"
                  properties:
                    system:
                      type: "string"
                      example: "SNOMED"
                datatypeNumber:
                  type: "object"
                  description: "object appearing only on Number datatypes"
                  properties:
                    maxValue:
                      type: "number"
                      format: "double"
                    minValue:
                      type: "number"
                      format: "double"
                    precision:
                      type: "number"
                      format: "double"
                datatypeText:
                  type: "object"
                  description: "object appearing only on Text datatypes"
                  properties:
                    maxLength:
                      type: "integer"
                      format: "int32"
                    minLength:
                      type: "integer"
                      format: "int32"
                    regex:
                      type: "string"
                      example: ".*:.*:.*"
                    rule:
                      type: "string"
                      example: ""
                    showAsTextArea:
                      type: "boolean"
                defaultAnswer:
                  type: "string"
                  example: "05/30/2014"
                editable:
                  type: "boolean"
                invisible:
                  type: "boolean"
                multiselect:
                  type: "boolean"
                required:
                  type: "boolean"
                unitsOfMeasure:
                  type: "array"
                  items:
                    type: "object"
                    properties:
                      code:
                        type: "string"
                        example: "m"
                      system:
                        type: "string"
                        example: "UCUM"
            repeat:
              type: "string"
              example: "5"
            section:
              type: "object"
            skipLogic:
              type: "object"
              properties:
                action:
                  type: "string"
                  example: "show"
                condition:
                  type: "string"
                  example: '"Starting Date" < 1/1/2015'
      designations:
        $ref: "#/definitions/Designations"
      definitions:
        $ref: "#/definitions/Definitions"
      classification:
        type: "array"
        items:
          $ref: "#/definitions/Classification"
      referenceDocuments:
        $ref: "#/definitions/ReferenceDocuments"
      displayProfiles:
        type: "array"
        items:
          type: "object"
          properties:
            displayInstructions:
              type: "boolean"
            displayInvisible:
              type: "boolean"
            displayNumbering:
              type: "boolean"
            displayType:
              type: "string"
              example: "Follow-up"
            displayValues:
              type: "boolean"
            name:
              type: "string"
              example: "print"
            numberOfColumns:
              type: "integer"
              format: "int32"
              example: "5"
            repeatFormat:
              type: "string"
              example: "#."
            sectionsAsMatrix:
              type: "boolean"
      properties:
        $ref: "#/definitions/Properties"
      ids:
        $ref: "#/definitions/Ids"
      attachments:
        $ref: "#/definitions/Attachments"
      history:
        $ref: "#/definitions/History"
  Attachments:
    type: "array"
    items:
      type: "object"
      properties:
        comment:
          type: "string"
          example: "Figure 1: Illustration"
        fileid:
          type: "string"
          example: "1"
        filename:
          type: "string"
          example: "1.jpeg"
        filesize:
          type: "integer"
          format: "int64"
        filetype:
          type: "string"
          example: "jpeg"
        isDefault:
          type: "boolean"
        pendingApproval:
          type: "boolean"
        scanned:
          type: "boolean"
        uploadedBy:
          $ref: "#/definitions/UserRef"
        uploadDate:
          type: "string"
          format: "date-time"
  Classification:
    type: "object"
    properties:
      elements:
        type: "array"
        items:
          $ref: "#/definitions/ClassificationElements"
      stewardOrg:
        type: "object"
        properties:
          name:
            type: "string"
            example: "NINDS"
      workingGroup:
        type: boolean
  ClassificationElements:
    type: "object"
    properties:
      elements:
        type: "array"
        items:
          $ref: "#/definitions/ClassificationElements"
      name:
        type: "string"
        example: "Disease"
  Comments:
    type: "array"
    items:
      type: "object"
      properties:
        created:
          type: "string"
          format: "date-time"
        element:
          type: "object"
          properties:
            eltId:
              type: "string"
              example: "views"
            eltType:
              type: "string"
              example: "cde"
        linkedTab:
          type: "string"
          example: "descriptionTab"
        pendingApproval:
          type: "boolean"
        replies:
          type: "array"
          items:
            type: "object"
            properties:
              created:
                type: "string"
                format: "date-time"
              pendingApproval:
                type: "boolean"
              status:
                type: "string"
                example: "resolved"
              text:
                type: "string"
                example: "update value"
              user:
                type: "string"
                example: "jdoe"
              usename:
                type: "string"
                example: "jdoe"
        status:
          type: "string"
          example: "resolved"
        text:
          type: "string"
          example: "update value"
        user:
          type: "string"
          example: "jdoe"
        usename:
          type: "string"
          example: "jdoe"
  CurationStatus:
    type: "string"
    enum:
      - "Qualified"
      - "Standard"
      - "Preferred Standard"
...
  Datatype:
    type: string
    enum:
      - "Date"
      - "Dynamic Code List"
      - "Externally Defined"
      - "File"
      - "Geo Location"
      - "Number"
      - "Text"
      - "Time"
      - "Value List"
  History:
    type: "array"
    items:
      type: "string"
      example: "57112ebf3245ae3033584c5a"
  Ids:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "string"
          example: "Qd4_a8FZqrX"
        source:
          type: "string"
          example: "NLM"
        version:
          type: "string"
          example: "3"
  Instructions:
    type: "object"
    properties:
      value:
        type: "string"
        example: "Part of the clinical trials."
      valueFormat:
        type: "string"
        example: "html"
  Designations:
    type: "array"
    items:
      type: "object"
      properties:
        designation:
          type: "string"
          example: "Title Name"
        source:
          type: "string"
          example: "NINDS"
        tags:
          type: "array"
          items:
            type: "object"
            properties:
              tag:
                type: "string"
                example: "Health"
  Definitions:
    type: "array"
    items:
      type: "object"
      properties:
        definition:
          type: "string"
          example: "This is a description of purpose"
        definitionFormat:
          type: "string"
          example: "html"
        source:
          type: "string"
          example: "NINDS"
        tags:
          type: "array"
          items:
            type: "object"
            properties:
              tag:
                type: "string"
                example: "Health"
  PermissibleValues:
    type: "array"
    items:
      type: "object"
      properties:
        codeSystemName:
          type: "string"
          example: "NCI Thesaurus"
        codeSystemVersion:
          type: "string"
          example: "1.0"
        permissibleValue:
          type: "string"
          example: "yes"
        valueMeaningCode:
          type: "string"
          example: "C46110"
        valueMeaningDefinition:
          type: "string"
          example: "affirmative declaration"
        valueMeaningName:
          type: "string"
          example: "Yes"
  Properties:
    type: "array"
    items:
      type: "object"
      properties:
        key:
          type: "string"
          example: "custom1"
        source:
          example: "NINDS"
          type: "string"
        value:
          type: "string"
          example: "10"
        valueFormat:
          type: "string"
          example: "html"
  ReferenceDocuments:
    type: "array"
    items:
      type: "object"
      properties:
        document:
          type: "string"
          example: "rd1"
        docType:
          type: "string"
          example: "rd1"
        languageCode:
          type: "string"
          example: "EN-US"
        providerOrg:
          type: "string"
          example: "NINDS"
        referenceDocumentId:
          type: "string"
          example: "ref1"
        source:
          type: "string"
          example: "NINDS"
        text:
          type: "string"
          example: "rd1"
        title:
          type: "string"
          example: "rd1"
        uri:
          type: "string"
          example: "http://www.ninds.com/125123u"
  RegistrationState:
    type: "object"
    properties:
      administrativeNote:
        type: "string"
        example: "approved"
      administrativeStatus:
        type: "string"
        example: "Nominal"
      effectiveDate:
        type: "string"
        format: "date-time"
      registrationStatus:
        type: "string"
        example: "Qualified"
      replacedBy:
        type: "object"
        properties:
          tinyId:
            type: "string"
            example: "Qd4_a8FZqrX"
      unresolvedIssue:
        type: "string"
        example: "none"
      untilDate:
        type: "string"
        format: "date-time"
...
  SearchDocumentResponseForm:
    type: object
    properties:
      resultsTotal:
        type: integer
        format: int32
      resultsRetrieved:
        type: integer
        format: int32
      from:
        type: integer
        format: int32
      docs:
        type: "array"
        items:
          $ref: "#/definitions/Form"
    required: [ "resultsTotal", "resultsRetrieved", "from", "docs" ]
  SearchApiOptions:
    type: object
    properties:
      nihEndorsed:
        description: "Endorsed by NIH"
        type: boolean
      excludeAllOrgs:
        description: "Exclude documents with more than 1 organization"
        type: boolean
      excludeOrgs:
        description: "Exclude document from specific organizations"
        type: array
        items:
          type: string
      page:
        description: "Specify page of result set, starting with 1"
        type: integer
      resultPerPage:
        description: "Number of results to retrieve at one time, maximum 100"
        type: integer
      searchTerm:
        description: "Search Query, see examples at https://cde.nlm.nih.gov/guides"
        type: string
      selectedDatatypes:
        description: "Data Elements with specific datatype"
        type: array
        items:
          $ref: "#/definitions/Datatype"
      selectedElements:
        description: "Classification"
        type: array
        items:
          type: string
      selectedElementsAlt:
        description: "Second Classification"
        type: array
        items:
          type: string
      selectedOrg:
        description: "Classifying Organization, not to be confused with steward"
        type: string
      selectedOrgAlt:
        description: "Second Classifying Organization, not to be confused with steward"
        type: string
      selectedStatuses:
        description: "Registration Status"
        type: array
        items:
          $ref: "#/definitions/CurationStatus"
        uniqueItems: true
  Source:
    type: "object"
    properties:
      sourceName:
        type: "string"
        example: "caDSR"
      created:
        type: "string"
        format: "date"
      updated:
        type: "string"
        format: "date"
      registrationStatus:
        type: "string"
        example: "Qualified"
      datatype:
        type: "string"
        example: "Number"
      copyright:
        type: "object"
        properties:
          value:
            type: "string"
            example: "MIT license"
          valueFormat:
            type: "string"
            example: "html"
  UserRef:
    type: "object"
    properties:
      username:
        type: "string"
        example: "jdoe"
      userId:
        type: "string"
        example: 'ObjectId("5711293936f82f59878e9c27")'
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"